[#508] Moved query plan to separate package
