+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/package-info.java
@@ -16,7 +16,7 @@
  */
 
 /**
- * Contains UDFs used by the Join query operators.
+ * Contains everything related to cypher query planning.
  */
 package org.gradoop.flink.model.impl.operators.matching.single.cypher;
