+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/planning/queryplan/BinaryNode.java
@@ -15,7 +15,7 @@
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.gradoop.flink.model.impl.operators.matching.single.cypher.planning;
+package org.gradoop.flink.model.impl.operators.matching.single.cypher.planning.queryplan;
 
 import java.util.Objects;
