+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/single/cypher/planning/queryplan/binary/package-info.java
@@ -16,7 +16,7 @@
  */
 
 /**
- * Contains all classes related to query plan representation.
+ * Contains binary plan node implementations.
  */
-package org.gradoop.flink.model.impl.operators.matching.single.cypher.planning;
+package org.gradoop.flink.model.impl.operators.matching.single.cypher.planning.queryplan.binary;
